+++ tests/testthat/test_PlotCatalogToPdf.DBS78Catalog.R
@@ -3,10 +3,9 @@ context("PlotCatalogToPdf.DBS78Catalog")
 test_that("PlotCatalogToPdf.DBS78Catalog function", {
   skip_if("" == system.file(package = "BSgenome.Hsapiens.1000genomes.hs37d5"))
   stopifnot(requireNamespace("BSgenome.Hsapiens.1000genomes.hs37d5"))
-  catalog.counts <- ReadCatalog("testdata/regress.cat.dbs.78.csv",
+  catalog.counts <- ReadCatalog("testdata/BTSG_WGS_PCAWG.dbs.78.csv",
                                 ref.genome = "GRCh37",
                                 region = "genome", catalog.type = "counts")
-  colnames(catalog.counts) <- paste0("HepG2_", 1 : 4)
   out <-
     PlotCatalogToPdf(catalog.counts, 
                      file = file.path(tempdir(), "PlotCatDBS78.counts.test.pdf"))